requirements.txt
Django==2.1.5
Pillow==5.4.1
django-pagedown==1.0.6
django-markdown-deux==1.0.5
markdown2==2.3.7
